requirements.txt
boto3
pyyaml
requests
psycopg2-binary
ddtrace
pandas